# docs/openapi/cluster-gateway.json
{
  "components": {
...
    "securitySchemes": {
      "gcoHmac": {
        "description": "Request-bound HMAC envelope minted by the API Gateway proxy Lambdas (`x-gco-signature-version`, `x-gco-timestamp`, `x-gco-nonce`, `x-gco-signature`): the signature covers the method, exact path and query, body digest, timestamp and nonce under the shared signing key from Secrets Manager. Verified by `gco/services/auth_middleware.py`; a failure is a 403. Original caller identity is authenticated upstream by API Gateway IAM.",
        "in": "header",
        "name": "x-gco-signature",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The HTTP surface of the internal ALB that fronts a GCO region: the Kubernetes Gateway `gco-system/gco-gateway` (GatewayClass `gco-aws-alb`, implemented by `gateway.k8s.aws/alb`, the AWS Load Balancer Controller) with one HTTPS:443 listener. Its HTTPRoute `gco-routes` sends each request to a Service by longest path prefix; every operation below is one a backend Service serves *and* the route delivers to it, tagged with that Service and annotated with the winning rule (`x-gco-route`). The ALB is `internal` and answers only the API Gateway proxy Lambdas (over Global Accelerator or from the regional bridge's VPC), which sign every request with the HMAC envelope (`x-gco-signature` and companions) that the services verify on every path except `/api/v1/health`, `/healthz`, `/metrics`, `/readyz`. The ALB terminates TLS with a deployment-local certificate and re-encrypts to a TLS proxy sidecar on each pod; the service documents describe the same operations as their applications see them.",
    "title": "GCO Cluster Gateway (internal ALB)",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cost/reports": {
      "get": {
        "description": "List this region's most recent cost report objects in S3.",
        "operationId": "list_cost_reports_api_v1_cost_reports_get",
        "parameters": [
          {
            "description": "List ad-hoc instead of scheduled reports",
            "in": "query",
            "name": "adhoc",
            "required": false,
            "schema": {
              "default": false,
              "description": "List ad-hoc instead of scheduled reports",
              "title": "Adhoc",
              "type": "boolean"
            }
          },
          {
            "description": "Maximum objects returned",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Maximum objects returned",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "List Cost Reports",
        "tags": [
          "manifest-processor",
          "Cost"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/",
          "service": "manifest-processor"
        }
      },
      "post": {
        "description": "Generate an ad-hoc cost report for the trailing window.",
        "operationId": "generate_cost_report_api_v1_cost_reports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "Generate Cost Report",
        "tags": [
          "manifest-processor",
          "Cost"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/",
          "service": "manifest-processor"
        }
      }
    },
    "/api/v1/cost/status": {
      "get": {
        "description": "Cost monitoring status for this region, including OpenCost health.",
        "operationId": "get_cost_status_api_v1_cost_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "Get Cost Status",
        "tags": [
          "manifest-processor",
          "Cost"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/",
          "service": "manifest-processor"
        }
      }
    },
...
    "/api/v1/jobs": {
      "delete": {
        "description": "Bulk delete jobs based on filters.",
        "operationId": "bulk_delete_jobs_api_v1_jobs_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "Bulk Delete Jobs",
        "tags": [
          "manifest-processor",
          "Jobs"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/",
          "service": "manifest-processor"
        }
      },
      "get": {
        "description": "List Kubernetes Jobs with pagination and filtering.",
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "description": "Filter by namespace",
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by namespace",
              "title": "Namespace"
            }
          },
          {
            "description": "Filter by status",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status",
              "title": "Status"
            }
          },
          {
            "description": "Maximum number of jobs to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Maximum number of jobs to return",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Number of jobs to skip",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of jobs to skip",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Sort field and order (field:asc|desc)",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "createdAt:desc",
              "description": "Sort field and order (field:asc|desc)",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated exact-match label filters (key=value only)",
            "in": "query",
            "name": "label_selector",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated exact-match label filters (key=value only)",
              "title": "Label Selector"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "List Jobs",
        "tags": [
          "manifest-processor",
          "Jobs"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/",
          "service": "manifest-processor"
        }
      }
    },
    "/api/v1/jobs/from-template/{name}": {
      "post": {
        "description": "Create a job from a template with parameter substitution.",
        "operationId": "create_job_from_template_api_v1_jobs_from_template__name__post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobFromTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "Create Job From Template",
        "tags": [
          "manifest-processor",
          "Templates"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/",
          "service": "manifest-processor"
        }
      }
    },
...
    "/api/v1/jobs/{namespace}/{name}/events": {
      "get": {
        "description": "Get events related to a Job.",
        "operationId": "get_job_events_api_v1_jobs__namespace___name__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "Get Job Events",
        "tags": [
          "manifest-processor",
          "Jobs"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/",
          "service": "manifest-processor"
        }
      }
    },
...
  },
  "servers": [
    {
      "description": "The region's internal ALB, registered in SSM as `/<project>/alb-hostname-<region>` by ga-registration and kept current by the health monitor; reachable from inside the VPC and from Global Accelerator only. Clients present SNI `backend.<project>.gco.internal`.",
      "url": "https://{alb_hostname}",
      "variables": {
...
    }
...
}